moved context.py code to corva/models/context.py
