+++ corva/app/scheduled.py
@@ -1,5 +1,5 @@
 from corva.app.base import BaseApp
-from corva.app.context import ScheduledContext
+from corva.models.context import ScheduledContext
 from corva.event.data.scheduled import ScheduledEventData
 from corva.event.event import Event
 from corva.event.loader.scheduled import ScheduledLoader
