+++ corva/app/task.py
@@ -2,7 +2,7 @@ import traceback
 from typing import Literal
 
 from corva.app.base import BaseApp
-from corva.app.context import TaskContext
+from corva.models.context import TaskContext
 from corva.models.task import TaskData, UpdateTaskInfoData
 from corva.event.event import Event
 from corva.event.loader.task import TaskLoader
